pyproject.toml
[project]
name = "freeai-utils"
version = "0.2.11"
description = "A free, zero-config AI utility toolkit with voice, text, image, and web integrations, pdf_docx readability"
authors = [
    { name="truongbaan", email="truongbaansoftware@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9.13"

dependencies = [
    "pyaudio",
    "numpy",
    "keyboard",
    "lameenc",
    "openai-whisper",
    "soundfile",
    "librosa",
    "torch",  # also needed for PIL use in transformers
    "google-generativeai",
    "pyperclip",
    "python-dotenv",
    "googlesearch-python",
    "requests",
    "beautifulsoup4",
    "readability-lxml",
    "transformers",
    "Pillow",
    "pyautogui",
    "easyocr",
    "gTTS",
    "playsound==1.2.2",  # required specific version
    "pyttsx3",
    "pymupdf",
    "pypdf",
    "pdfplumber",
    "python-docx",
    "mutagen",
    "haystack-ai"
]

[project.urls]
Homepage = "https://github.com/truongbaan/Utility-python-library"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["freeai_utils"]

[tool.setuptools.package-data]
"*" = ["README.md", "API.md", "THIRD_PARTY.md"]

[project.scripts]
freeai-utils-setup = "freeai_utils._setup:install_default_model"
